pyproject.toml
[project]
name = "murlix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "google-adk",
    "python-dotenv>=1.1.1",
    "questionary>=2.1.0",
    "rich>=14.0.0",
]

[tool.uv.sources]
google-adk = { git = "https://github.com/google/adk-python.git", rev = "main" }
